add a different innocuous error
If dependencies are waiting, then they will definitely trigger the alert
as they are not reconciling at all while they wait for dependencies.

So to verify https://github.com/fluxcd/flux2/discussions/2942 we need to
confirm that Flux reports a steady 1.0 when a Kustomization with no
dependency does try to reconcile, but fails to apply. That Kustomization
was never ready and so should consistently report Failure on the metric.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>